JAMES-3534 Dedicated class for MemoryCustomIdentityDAO
